Add files via upload
Natural Language Processing, is a subfield of artificial intelligence that focuses on the interaction between computers and human language. It involves the development of algorithms and models that enable computers to understand, interpret, and generate human language in a way that is meaningful and useful.

The goal of NLP is to bridge the gap between human language and computer systems, allowing machines to process and understand natural language input from humans. This involves a range of tasks, including speech recognition, natural language understanding, language generation, sentiment analysis, and machine translation.

One of the key challenges in NLP is the inherent complexity and ambiguity of human language. Language is dynamic, context-dependent, and often contains multiple layers of meaning. NLP algorithms and models must be able to handle these complexities and accurately interpret the intended meaning of a given text or speech.

NLP techniques typically involve the use of statistical models, machine learning algorithms, and linguistic rules. These models are trained on large datasets of annotated language examples, allowing them to learn patterns and relationships between words, phrases, and sentences. Some popular NLP techniques include named entity recognition, part-of-speech tagging, syntactic parsing, and semantic analysis.