Add files via upload
Topic Modeling and Sentiment Analysis are two important techniques in the field of Natural Language Processing (NLP) that play a crucial role in understanding and extracting insights from textual data.
One popular algorithm for topic modeling is Latent Dirichlet Allocation (LDA). LDA assumes that each document is a mixture of a small number of topics, and each topic is a probability distribution over words. By applying LDA to a collection of documents, we can uncover the underlying topics and the distribution of these topics across the documents.